requirements.txt
# FlowDiff - Function Call Tree Visualizer Dependencies

# Web server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# CLI framework
typer==0.9.0
rich==13.7.0

# Configuration
pyyaml>=6.0

# LLM integration (optional - for entry point filtering)
anthropic>=0.39.0
